requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
databricks-sdk>=0.81.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
psycopg[binary,pool]>=3.0
cachetools>=5.3
